optimize: renamed generator to actuator
