+++ runtime/actuator/actuator_test.go
@@ -1,4 +1,4 @@
-package generator
+package actuator
 
 import (
 	"context"
