+++ runtime/actuator/errors.go
@@ -1,4 +1,4 @@
-package generator
+package actuator
 
 import (
 	"errors"
